Added image and image view creation for the swap chain
Next, add in system to present an image.

+++ VEngine/Win32Window.cpp
@@ -1,7 +1,7 @@
 #include "Win32Window.h"
 
 
-Win32Window::Win32Window(std::string WindowName, float xResolution, float yResolution, bool isFullscreen)
+Win32Window::Win32Window(std::string WindowName, int xResolution, int yResolution, bool isFullscreen)
 {
 	windowData.connection = GetModuleHandle(NULL);
